issue-34: move operation in separate file
