+++ wsdl-parser/src/parser/mod.rs
@@ -2,6 +2,7 @@ use crate::parser::types::Definitions;
 
 pub mod types;
 mod constants;
+pub mod operation;
 
 pub fn parse(text: &str)  {
     let doc = roxmltree::Document::parse(&text).expect("Parse document error");
@@ -35,6 +36,50 @@ pub fn parse(text: &str)  {
     //     println!("{:#?}", i);
     // }
 }
+
+
+pub trait WsdlElement {
+    fn wsdl_type(&self) -> ElementType;
+}
+
+#[derive(Debug, PartialEq)]
+pub enum ElementType {
+    Binding,
+    Definitions,
+    Documentation,
+    Import,
+    Input,
+    Fault,
+    Message,
+    Operation,
+    Output,
+    Part,
+    PortType,
+    Types,
+    UnknownElement(String)
+}
+
+impl<'a> WsdlElement for roxmltree::Node<'a, '_> {
+    fn wsdl_type(&self) -> ElementType {
+        use ElementType::*;
+        match self.tag_name().name() {
+            "binding" => Binding,
+            "definitions" => Definitions,
+            "documentation" => Documentation,
+            "import" => Import,
+            "input" => Input,
+            "fault" => Fault,
+            "message" => Message,
+            "operation" => Operation,
+            "output" => Output,
+            "part" => Part,
+            "portType" => PortType,
+            "types" => Types,
+            _ => UnknownElement(self.tag_name().name().to_string()),
+        }
+    }
+}
+
